+++ tests/testthat/test-writeCounts.R
@@ -21,11 +21,11 @@ test_that("writeCounts", {
 test_that("Require a matrix, and don't allow data frames.", {
     expect_error(
         object = writeCounts(mtcars),
-        regexp = "mtcars is not a matrix"
+        regexp = "'mtcars' is not a matrix"
     )
 })
 
-test_that("Check that `eval_bare` call errors on missing object.", {
+test_that("Check that 'eval_bare()' call errors on missing object.", {
     expect_error(
         object = writeCounts(XXX),
         regexp = "object 'XXX' not found"